ensure that observedSet resets observed objects after changes
this was a pretty major oversight on my part, and unfortunately, the tests are structure such that the "organic" callback of Object.observe is never depended upon.

What really needs to be done here is to make most of the tests be async using the then() pattern, but this bug is blocking polymer release, so I'm going to do that work in a follow-on patch.

https://github.com/Polymer/observe-js/issues/42
R=arv
BUG=

Review URL: https://codereview.appspot.com/53490043